Rename bridgeApi.js to EdgeProvider.js
